+++ src/db/schema/volunteer/volunteer-opportunities.ts
@@ -26,6 +26,7 @@ export const volunteerOpportunityStatus = pgEnum("volunteer_opportunity_status",
 
 export const volunteerApplicationStatus = pgEnum("volunteer_application_status", [
   "SUBMITTED",
+  "APPROVAL",
   "ACCEPTED",
   "REJECTED",
   "WITHDRAWN",
@@ -226,7 +227,7 @@ export const volunteerApplication = pgTable(
     uniqueIndex("volunteer_application_applicant_role_active_unique_idx")
       .on(table.applicantId, table.roleId)
       .where(
-        sql`${table.status} in ('SUBMITTED', 'ACCEPTED')`,
+        sql`${table.status} in ('SUBMITTED', 'APPROVAL', 'ACCEPTED')`,
       ),
     index("volunteer_application_opportunity_idx").using(
       "btree",
