+++ src/modules/forum/lib/constants.ts
@@ -1,5 +1,6 @@
-export const FORUM_UUID_RE =
-  /^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+import { UUID_RE } from "../../../lib/constant";
+
+export const FORUM_UUID_RE = UUID_RE;
 
 export const FORUM_ADVISORY_LOCK_NAMESPACE = 100;
 export const FORUM_CATEGORY_DISPLAY_ORDER_LOCK_KEY = 1;
