+++ src/routes/routes.ts
@@ -1,5 +1,6 @@
 import { OpenAPIHono } from "@hono/zod-openapi";
 import type { AppBindings } from "../lib/types";
+import { applicationRouter } from "../modules/application/application.router";
 import { authRouter } from "../modules/auth/auth.router";
 import { answersRouter } from "../modules/forum/answers/answers.router";
 import { categoriesRouter } from "../modules/forum/categories/categories.router";
@@ -27,6 +28,7 @@ routes.route("/forum/questions", questionsRouter);
 routes.route("/forum/answer", answersRouter);
 routes.route("/volunteer", postVolunteerRouter);
 routes.route("/volunteer/public", publicPostVolunteerRouter);
+routes.route("/volunteer/applications", applicationRouter);
 routes.route("/forum/public/questions", publicQuestionsRouter);
 routes.route("/forum/public/category", publicCategoriesRouter);
 routes.route("/forum/public/answer", publicAnswersRouter);
